Extract common AuthSchemeInterceptor logic into a base class
